refactor(react-tree-grid): consolidate threaded story coverage
